+++ Source/Bifrost.Web.Specs/Services/for_RestServiceMethodInvoker/given/a_rest_service_method_invoker_and_a_service_call.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using System.Collections.Specialized;
+using System.Collections.Generic;
 using Machine.Specifications;
 
 namespace Bifrost.Web.Specs.Services.for_RestServiceMethodInvoker.given
@@ -9,13 +9,13 @@ namespace Bifrost.Web.Specs.Services.for_RestServiceMethodInvoker.given
         protected const string base_url = "ServiceWithMethods";
         protected static ServiceWithMethods service_instance;
         protected static Uri uri;
-        protected static NameValueCollection parameters;
+        protected static IDictionary<string, string> parameters;
 
         Establish context = () =>
         {
             service_instance = new ServiceWithMethods();
-            uri = new Uri(string.Format("http://localhost/{0}/{1}", base_url, ServiceWithMethods.ComplexInputNoOutputMethod));
-            parameters = new NameValueCollection();
+            uri = new Uri($"http://localhost/{base_url}/{ServiceWithMethods.ComplexInputNoOutputMethod}");
+            parameters = new Dictionary<string, string>();
         };
     }
 }